docker-compose-files/jaeger/graphql-api-python/pyproject.toml
[project]
name = "jager-python"
version = "0.1.0"
description = "GraphQL Python app with Jaeger tracing and OpenTelemetry"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strawberry-graphql[fastapi]>=0.245.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-exporter-otlp-proto-http>=1.28.0",
    "python-json-logger>=3.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = ["ruff>=0.8.0", "mypy>=1.13.0", "pre-commit>=4.0.0"]

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    ".venv",
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade (Python version upgrades)
    "B",   # flake8-bugbear (common bugs)
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = []
fixable = ["ALL"] # Allow all auto-fixable rules to be fixed
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
